Added camera tracking
Camera now tracks the ball. Meddled a little with the code from the
course as I felt the camera was ending too close to the head pin and at
the moment the collisions with pins aren't helping. May revert back
later.

Also changed initial ball speed to a Vector3 initial velocity.