Fixed issue with fast-moving objects and collisions
Changed the collision detection settings for the Ball and the Bowling
Pin to allow for fast-moving objects.

This problem was found when using the swipe functionality so I'm deeming
it a fix for now until we look at perhaps clamping the velocity to a
suitable value later.

Noting that these changes can severely impact Physics performance so may
need to look back at this later.