Added Pin script to check if Pin is standing
A slight alteration from the course's way of checking if a pin is
upright. Keeping with the threshold of angles but providing a more
accurate angle using the Vector3.Angle() function with the world Up
vector and the Pin's up vector, rather than checking the threshold
against the rotation in X and Z separately.