Many more tests, much more secure ActionMaster
Really nailing down on adding a lot of tests, particularly those
relating to the last frame in a game of bowling and whether we need to
Tidy the pins, Reset the pins or End the Game. As the ActionMaster code
itself shows, the main issue is surrounding the 20th Bowl when either of
the 3 conditions can occur for different reasons.

Also fixing a bug with a 0,10 spare in a frame causing the next less
than 10 bowl to be considered the End of a Turn rather than a Tidy.